adding vmanage context
