+++ nornir_nautobot/utils/helpers.py
@@ -3,10 +3,7 @@
 import errno
 import importlib
 import logging
-import os
-import re
 import traceback
-import unicodedata
 from typing import Any
 
 from nornir_nautobot.constants import ERROR_CODES
@@ -16,10 +13,10 @@ LOGGER = logging.getLogger(__name__)
 
 def make_folder(folder):
     """Helper method to sanely create folders."""
-    if not os.path.exists(folder):
+    if not pathlib.Path(folder).exists():
         # Still try and except, since their may be race conditions.
         try:
-            os.makedirs(folder)
+            pathlib.Path(folder).mkdir(parents=True)
         except OSError as error:
             if error.errno != errno.EEXIST:
                 raise
@@ -84,40 +81,3 @@ def get_error_message(error_code: str, **kwargs: Any) -> str:
     except Exception:  # pylint: disable=broad-except
         error_message = "Error Code was found, but failed to format message, unknown cause."
     return f"{error_code}: {error_message}"
-
-
-def command_to_filename(command, replacement="_"):
-    """
-    Convert a command string into a filesystem-safe filename.
-
-    This function sanitizes a command string so it can safely be used as a filename by:
-    1. Normalizing Unicode characters to their ASCII equivalents.
-    2. Replacing the pipe symbol '|' (with or without surrounding spaces) with two replacement characters.
-    3. Replacing characters that are illegal in most filesystems (e.g., / : * ? " < >) with the replacement character.
-    4. Replacing all spaces with the replacement character.
-
-    Args:
-        command (str): The input command string to sanitize.
-        replacement (str): The character to use as a substitute for illegal or special characters. Default is underscore ('_').
-
-    Returns:
-        str: A sanitized, ASCII-only, filesystem-safe version of the command string suitable for use as a filename.
-    """
-    # 1. Normalize Unicode characters (e.g., accented characters to ASCII equivalents)
-    #    and then ignore non-ASCII. This handles many international characters.
-    command_file_name = unicodedata.normalize("NFKD", command).encode("ascii", "ignore").decode("ascii")
-
-    # 2. Replace the '|' with the replacement character
-    command_file_name = re.sub(r"\s*\|\s*", 2 * replacement, command_file_name)
-
-    # 3. Replace characters illegal in most filesystems with the replacement character
-    #    Common illegal characters: / \ : * ? " < > |
-    #    Also replace leading/trailing whitespace, and control characters.
-    #    This regex targets common illegal filename characters and also handles multiple
-    #    replacement characters in a row (e.g., 'foo///bar' becomes 'foo_bar').
-    command_file_name = re.sub(r"\s*[\/\\:*?\"<>]\s*", replacement, command_file_name)
-
-    # 4. Replace spaces with the replacement character
-    command_file_name = re.sub(r"\s+", replacement, command_file_name).strip(replacement)
-
-    return command_file_name
